--- a/addons/missileguidance/ACE_Settings.hpp
+++ b/addons/missileguidance/ACE_Settings.hpp
@@ -2,7 +2,6 @@ class ACE_Settings {
     class GVAR(enabled) {
         value = 2;
         typeName = "SCALAR";
-        isClientSettable = 1;
         displayName = CSTRING(Title);
         description = CSTRING(Desc);
         values[] = {CSTRING(Off), CSTRING(PlayerOnly), CSTRING(PlayerAndAi)};