--- a/addons/missileguidance/XEH_PREP.hpp
+++ b/addons/missileguidance/XEH_PREP.hpp
@@ -1,6 +1,4 @@
 
-PREP(rotateVectLineGetMap);
-PREP(rotateVectLine);
 PREP(changeMissileDirection);
 
 PREP(checkSeekerAngle);
@@ -17,11 +15,11 @@ PREP(doHandoff);
 PREP(handleHandoff);
 
 // Attack Profiles
-PREP(attackProfile_LIN);
-PREP(attackProfile_DIR);
-PREP(attackProfile_MID);
-PREP(attackProfile_HI);
 PREP(attackProfile_AIR);
+PREP(attackProfile_DIR);
+PREP(attackProfile_HI);
+PREP(attackProfile_LIN);
+PREP(attackProfile_MID);
 
 // Javelin profiles
 PREP(attackProfile_JAV_DIR);