--- a/addons/missileguidance/functions/fnc_attackProfile_LIN.sqf
+++ b/addons/missileguidance/functions/fnc_attackProfile_LIN.sqf
@@ -22,6 +22,8 @@ params ["_seekerTargetPos", "_args"];
 _args params ["_firedEH"];
 _firedEH params ["_shooter","","","","","","_projectile"];
 
+if (_seekerTargetPos isEqualTo [0,0,0]) exitWith {_seekerTargetPos};
+
 private _shooterPos = getPosASL _shooter;
 private _projectilePos = getPosASL _projectile;