--- a/addons/missileguidance/functions/fnc_checkSeekerAngle.sqf
+++ b/addons/missileguidance/functions/fnc_checkSeekerAngle.sqf
@@ -1,31 +1,35 @@
 /*
  * Author: jaynus
- * Returns whether the target position is within the maximum angle FOV of the provided seeker 
+ * Returns whether the target position is within the maximum angle FOV of the provided seeker
  * objects current direction.
  *
  * Arguments:
- * 0: Seeker [Object]
- * 1: Target [Position]
- * 2: Max Angle [Degrees]
- * 
+ * 0: Seeker <OBJECT>
+ * 1: Target PosASL <ARRAY>
+ * 2: Max Angle (degrees) <NUMBER>
+ *
  * Return Value:
- *    Boolean
+ * Can See <BOOL>
+ *
+ * Example:
+ * [player, cursorTarget, 45] call ace_missileguidance_fnc_checkSeekerAngle;
+ *
+ * Public: No
  */
-
+// #define DEBUG_MODE_FULL
 #include "script_component.hpp"
-private ["_seeker", "_targetPos", "_seekerMaxAngle", "_sensorPos", "_testPointVector", "_testDotProduct"];
 
-_seeker = _this select 0;
-_targetPos = _this select 1;
-_seekerMaxAngle = _this select 2;
+params ["_seeker", "_targetPos", "_seekerMaxAngle"];
 
-_sensorPos = getPosASL _seeker;
+private _sensorPos = getPosASL _seeker;
 
-_testPointVector = vectorNormalized (_targetPos vectorDiff _sensorPos);
-_testDotProduct = (vectorNormalized (velocity _seeker)) vectorDotProduct _testPointVector;
+private _testPointVector = vectorNormalized (_targetPos vectorDiff _sensorPos);
+private _testDotProduct = (vectorNormalized (velocity _seeker)) vectorDotProduct _testPointVector;
 
-if(_testDotProduct < (cos _seekerMaxAngle)) exitWith {
-    false 
+TRACE_2("fov",acos _testDotProduct,_seekerMaxAngle);
+
+if (_testDotProduct < (cos _seekerMaxAngle)) exitWith {
+    false
 };
 
 true