--- a/addons/missileguidance/functions/fnc_doHandoff.sqf
+++ b/addons/missileguidance/functions/fnc_doHandoff.sqf
@@ -1,3 +1,4 @@
+// Not currently used
 #include "script_component.hpp"
 PARAMS_2(_target,_args);