… addons/missileguidance/functions/fnc_guidancePFH.sqf
… addons/missileguidance/functions/fnc_handleHandoff.sqf
… addons/missileguidance/functions/fnc_onFired.sqf
… addons/missileguidance/functions/fnc_onIncomingMissile.sqf
… addons/missileguidance/functions/fnc_rotateVectLine.sqf
… addons/missileguidance/functions/fnc_rotateVectLineGetMap.sqf
… addons/missileguidance/functions/fnc_seekerType_Optic.sqf
… addons/missileguidance/functions/fnc_seekerType_SALH.sqf
… addons/missileguidance/script_component.hpp