- Headers
- Modern params/private
- Moved laser seeker range check to ace_laser (so we don't waste time on lasers we can't see)
- Added `seekLastTargetPos` config (enabled for javelin). If seeker loses LOS of target, will still continue to last known position.
- If seeker returns no position, missile will fly straight instead of trying an attack profile (useful for DAGR)

Picked up some performance gains on the main guidance PFEH:
Before: `Average: 0.55397s / 1222 = 0.453331ms`
After: `Average: 0.436127s / 1269 = 0.343677ms`